+++ include/simulation.h
@@ -28,22 +28,22 @@
 
 __global__ void simulate_worker_kernel(
     // Neural net
-    const float* nn_params,
+    const double* nn_params,
     const int* nn_layer_sizes,
     int nn_num_layers,
-    float greedy_cost_prob,
+    double greedy_cost_prob,
     // Worker state. inventory_scratch is the full [n_products, n_nodes_plus1]
     // inventory (a fresh per-iteration copy). Products are partitioned 1:1 to
     // workers, so each worker only ever touches rows for its own products — the
     // writes are disjoint across workers, no per-worker private copy needed (P13).
-    float* inventory_scratch,   // [n_products, n_nodes_plus1]
-    const float* init_capacity, // [n_nodes_plus1]
+    double* inventory_scratch,   // [n_products, n_nodes_plus1]
+    const double* init_capacity, // [n_nodes_plus1]
     // Events for this iteration
     const int* events_product,           // [n_workers, num_steps] — original product id
     const int* events_quantity,          // [n_workers, num_steps]
     const int* event_ntf_full,           // [n_events, n_nodes_plus1] — immutable, device-resident
     const int* order_2D_index,           // [n_workers, num_steps] — global event idx per slot (-1 if none)
-    const float* events_capacity_delta,  // [n_workers, num_steps, n_nodes_plus1]
+    const double* events_capacity_delta,  // [n_workers, num_steps, n_nodes_plus1]
     const int* valid_mask,               // [n_workers, num_steps]
     const uint64_t* rng_keys,            // [n_workers]
     // Dimensions
@@ -61,15 +61,15 @@ __global__ void simulate_worker_kernel(
     if (threadIdx.x != 0) return;
 
     // Local capacity state (copy from init_capacity)
-    float local_capacity[MAX_NODES];
+    double local_capacity[MAX_NODES];
     for (int n = 0; n < n_nodes_plus1; n++) {
         local_capacity[n] = init_capacity[n];
     }
 
     // Scratch buffers for NN inference
-    float scratch_a[MAX_NN_DIM];
-    float scratch_b[MAX_NN_DIM];
-    float nn_output[MAX_NN_DIM];
+    double scratch_a[MAX_NN_DIM];
+    double scratch_b[MAX_NN_DIM];
+    double nn_output[MAX_NN_DIM];
 
     uint64_t key = rng_keys[worker_id];
 
@@ -82,7 +82,7 @@ __global__ void simulate_worker_kernel(
         }
 
         // Apply capacity delta from other threads
-        const float* cap_delta = events_capacity_delta + (size_t)flat_idx * n_nodes_plus1;
+        const double* cap_delta = events_capacity_delta + (size_t)flat_idx * n_nodes_plus1;
         for (int n = 0; n < n_nodes_plus1; n++) {
             local_capacity[n] -= cap_delta[n];
         }
@@ -94,7 +94,7 @@ __global__ void simulate_worker_kernel(
         const int* node_ntf = event_ntf_full + (size_t)order_2D_index[flat_idx] * n_nodes_plus1;
 
         // Inventory row for this product (rows are disjoint across workers).
-        float* inv_row = inventory_scratch + (size_t)product * n_nodes_plus1;
+        double* inv_row = inventory_scratch + (size_t)product * n_nodes_plus1;
 
         // Split key
         uint64_t subkey;
@@ -111,8 +111,8 @@ __global__ void simulate_worker_kernel(
         fulfill_2D[flat_idx] = node;
 
         // Update inventory and capacity
-        inv_row[node] -= (float)quantity;
-        local_capacity[node] -= (float)quantity;
+        inv_row[node] -= (double)quantity;
+        local_capacity[node] -= (double)quantity;
     }
 }
 
@@ -123,33 +123,33 @@ __global__ void simulate_worker_kernel(
 // One WARP (32 lanes) processes one worker. The step loop stays serial (state is
 // carried across steps), but within each event the expensive MLP matvec is split
 // across lanes, and the small norm/logsumexp reductions are done with warp
-// shuffles. The warp reductions sum floats in a different order than the
-// single-thread kernel, so the output is NOT bit-identical — it is verified
-// against the sequential scan with a node-decision match tolerance instead.
+// shuffles. The warp reductions sum in a different order than the single-thread
+// kernel, so the output is NOT bit-identical — it is verified against the
+// sequential scan with a node-decision match tolerance instead.
 
 #define WARPS_PER_BLOCK 8   // 256 threads/block
 
-__device__ inline float warp_reduce_sum(float v) {
+__device__ inline double warp_reduce_sum(double v) {
     for (int o = 16; o > 0; o >>= 1) v += __shfl_down_sync(0xffffffffu, v, o);
     return v;  // lane 0 holds the total
 }
-__device__ inline float warp_reduce_max(float v) {
-    for (int o = 16; o > 0; o >>= 1) v = fmaxf(v, __shfl_down_sync(0xffffffffu, v, o));
+__device__ inline double warp_reduce_max(double v) {
+    for (int o = 16; o > 0; o >>= 1) v = fmax(v, __shfl_down_sync(0xffffffffu, v, o));
     return v;  // lane 0 holds the max
 }
 
 __global__ void simulate_worker_kernel_warp(
-    const float* nn_params,
+    const double* nn_params,
     const int* nn_layer_sizes,
     int nn_num_layers,
-    float greedy_cost_prob,
-    float* inventory_scratch,            // [n_products, n_nodes_plus1] — disjoint rows per worker (P13)
-    const float* init_capacity,
+    double greedy_cost_prob,
+    double* inventory_scratch,            // [n_products, n_nodes_plus1] — disjoint rows per worker (P13)
+    const double* init_capacity,
     const int* events_product,           // original product id per slot
     const int* events_quantity,
     const int* event_ntf_full,           // [n_events, n_nodes_plus1] — immutable, device-resident
     const int* order_2D_index,           // [n_workers, num_steps] — global event idx per slot
-    const float* events_capacity_delta,
+    const double* events_capacity_delta,
     const int* valid_mask,
     const uint64_t* rng_keys,
     int n_workers,
@@ -164,13 +164,13 @@ __global__ void simulate_worker_kernel_warp(
     if (worker_id >= n_workers) return;  // uniform across the warp
 
     // Per-warp shared state and NN scratch (ping-pong buffers a/b, output, cap).
-    __shared__ float s_cap[WARPS_PER_BLOCK][MAX_NODES];
-    __shared__ float s_a[WARPS_PER_BLOCK][MAX_NN_DIM];
-    __shared__ float s_b[WARPS_PER_BLOCK][MAX_NN_DIM];
-    __shared__ float s_out[WARPS_PER_BLOCK][MAX_NN_DIM];
+    __shared__ double s_cap[WARPS_PER_BLOCK][MAX_NODES];
+    __shared__ double s_a[WARPS_PER_BLOCK][MAX_NN_DIM];
+    __shared__ double s_b[WARPS_PER_BLOCK][MAX_NN_DIM];
+    __shared__ double s_out[WARPS_PER_BLOCK][MAX_NN_DIM];
     __shared__ int   s_node[WARPS_PER_BLOCK];
 
-    float* cap = s_cap[wpb];
+    double* cap = s_cap[wpb];
     for (int n = lane; n < n_nodes_plus1; n += 32) cap[n] = init_capacity[n];
     __syncwarp();
 
@@ -187,7 +187,7 @@ __global__ void simulate_worker_kernel_warp(
             continue;
         }
 
-        const float* cap_delta = events_capacity_delta + (size_t)flat_idx * n_nodes_plus1;
+        const double* cap_delta = events_capacity_delta + (size_t)flat_idx * n_nodes_plus1;
         for (int n = lane; n < n_nodes_plus1; n += 32) cap[n] -= cap_delta[n];
         __syncwarp();
 
@@ -196,7 +196,7 @@ __global__ void simulate_worker_kernel_warp(
         // Index the immutable, device-resident full ntf by this slot's global event
         // index (valid here: the !valid_mask branch above already `continue`d).
         const int* node_ntf = event_ntf_full + (size_t)order_2D_index[flat_idx] * n_nodes_plus1;
-        float* inv_row = inventory_scratch + (size_t)product * n_nodes_plus1;
+        double* inv_row = inventory_scratch + (size_t)product * n_nodes_plus1;
 
         uint64_t subkey;
         split_key(key, key, subkey);                 // identical on every lane
@@ -204,8 +204,8 @@ __global__ void simulate_worker_kernel_warp(
 
         if (use_cost) {
             // --- MLP forward pass, outputs split across lanes ---
-            float* a = s_a[wpb];
-            float* b = s_b[wpb];
+            double* a = s_a[wpb];
+            double* b = s_b[wpb];
             for (int i = lane; i < in_size; i += 32) a[i] = inv_row[i];
             __syncwarp();
 
@@ -213,51 +213,51 @@ __global__ void simulate_worker_kernel_warp(
             for (int layer = 0; layer < nn_num_layers; layer++) {
                 int nin  = nn_layer_sizes[layer];
                 int nout = nn_layer_sizes[layer + 1];
-                const float* W = nn_params + param_offset;
-                const float* bias = W + nout * nin;
+                const double* W = nn_params + param_offset;
+                const double* bias = W + nout * nin;
                 param_offset += nout * nin + nout;
                 bool last = (layer == nn_num_layers - 1);
-                float* dst = last ? s_out[wpb] : b;
+                double* dst = last ? s_out[wpb] : b;
                 // Each lane computes a disjoint set of outputs; the inner dot is
                 // serial per output, so its summation order matches the scalar
                 // kernel exactly (only the later reductions differ).
                 for (int i = lane; i < nout; i += 32) {
-                    float sum = bias[i];
+                    double sum = bias[i];
                     for (int j = 0; j < nin; j++) sum += W[i * nin + j] * a[j];
-                    dst[i] = last ? sum : fmaxf(0.0f, sum);
+                    dst[i] = last ? sum : fmax(0.0, sum);
                 }
                 __syncwarp();
-                if (!last) { float* t = a; a = b; b = t; }  // swap ping-pong
+                if (!last) { double* t = a; a = b; b = t; }  // swap ping-pong
             }
 
             // --- logsumexp normalize over s_out[0..out_size) (warp reductions) ---
-            float* out = s_out[wpb];
-            float m = -1e30f;
-            for (int i = lane; i < out_size; i += 32) m = fmaxf(m, out[i]);
+            double* out = s_out[wpb];
+            double m = -1e30;
+            for (int i = lane; i < out_size; i += 32) m = fmax(m, out[i]);
             m = warp_reduce_max(m);
             m = __shfl_sync(0xffffffffu, m, 0);
-            float se = 0.0f;
-            for (int i = lane; i < out_size; i += 32) se += expf(out[i] - m);
+            double se = 0.0;
+            for (int i = lane; i < out_size; i += 32) se += exp(out[i] - m);
             se = warp_reduce_sum(se);
             se = __shfl_sync(0xffffffffu, se, 0);
-            float lse = m + logf(se);
+            double lse = m + log(se);
             for (int i = lane; i < out_size; i += 32) out[i] -= lse;
             __syncwarp();
 
             // --- norm of the (normalized) output ---
-            float nr = 0.0f;
+            double nr = 0.0;
             for (int i = lane; i < n_nodes_plus1; i += 32) nr += out[i] * out[i];
             nr = warp_reduce_sum(nr);
             nr = __shfl_sync(0xffffffffu, nr, 0);
-            nr = sqrtf(nr + 1e-12f);
+            nr = sqrt(nr + 1e-12);
 
             // --- feasibility scan (cheap, serial on lane 0) ---
             if (lane == 0) {
                 int chosen = node_ntf[n_nodes_plus1 - 1];
                 for (int rank = 0; rank < n_nodes_plus1; rank++) {
                     int node = node_ntf[rank];
-                    float iv = inv_row[node] + roundf(1e-7f * out[rank] / nr);
-                    if (cap[node] >= (float)quantity && iv >= (float)quantity) {
+                    double iv = inv_row[node] + round(1e-7 * out[rank] / nr);
+                    if (cap[node] >= (double)quantity && iv >= (double)quantity) {
                         chosen = node; break;
                     }
                 }
@@ -266,10 +266,10 @@ __global__ void simulate_worker_kernel_warp(
         } else {
             // greedy_capacity: argmax feasible capacity (serial on lane 0)
             if (lane == 0) {
-                int best_node = -1; float best_cap = -1.0f;
+                int best_node = -1; double best_cap = -1.0;
                 for (int rank = 0; rank < n_nodes_plus1 - 1; rank++) {
                     int node = node_ntf[rank];
-                    if (cap[node] >= (float)quantity && inv_row[node] >= (float)quantity
+                    if (cap[node] >= (double)quantity && inv_row[node] >= (double)quantity
                         && cap[node] > best_cap) {
                         best_cap = cap[node]; best_node = node;
                     }
@@ -282,8 +282,8 @@ __global__ void simulate_worker_kernel_warp(
 
         if (lane == 0) {
             fulfill_2D[flat_idx] = node;
-            inv_row[node] -= (float)quantity;
-            cap[node] -= (float)quantity;
+            inv_row[node] -= (double)quantity;
+            cap[node] -= (double)quantity;
         }
         __syncwarp();
     }
@@ -297,7 +297,7 @@ __global__ void simulate_worker_kernel_warp(
 // largest per-window structure (evc·np1) and was both computed on the host (a
 // 31M-element cumsum + gather, the steady-state prep bottleneck) and uploaded H2D
 // (124 MB). Quantities are all 1, so every cumsum partial is an exact integer
-// (|·| ≤ eff ≪ 2^24) — float adds are exact regardless of order — so the device
+// (|·| ≤ eff ≪ 2^24) — double adds are exact regardless of order — so the device
 // result is **bit-identical** to the host version.
 
 // Per-column prefix sum: cumsum[t,n] = -count(s<=t : fulfill[s]==n) (×quantity).
@@ -311,7 +311,7 @@ __global__ void capacity_cumsum_kernel(
     const int* __restrict__ fulfill,    // [eff]
     const int* __restrict__ quant,      // [eff]
     int eff, int np1,
-    float* __restrict__ cumsum          // [eff, np1]
+    double* __restrict__ cumsum         // [eff, np1]
 ) {
     const int n = blockIdx.x;
     if (n >= np1) return;
@@ -319,26 +319,26 @@ __global__ void capacity_cumsum_kernel(
     const int lane = threadIdx.x & 31;
     const int warp = threadIdx.x >> 5;
     const int nwarps = T >> 5;
-    __shared__ float wsum[32];   // per-warp totals
-    __shared__ float wexcl[32];  // exclusive prefix of the warp totals
-    __shared__ float carry_s;    // running column total across chunks
-    __shared__ float btot_s;     // this chunk's block total
-    if (threadIdx.x == 0) carry_s = 0.0f;
+    __shared__ double wsum[32];   // per-warp totals
+    __shared__ double wexcl[32];  // exclusive prefix of the warp totals
+    __shared__ double carry_s;    // running column total across chunks
+    __shared__ double btot_s;     // this chunk's block total
+    if (threadIdx.x == 0) carry_s = 0.0;
     __syncthreads();
 
     for (int base = 0; base < eff; base += T) {
         int t = base + threadIdx.x;
-        float x = (t < eff && fulfill[t] == n) ? -(float)quant[t] : 0.0f;
+        double x = (t < eff && fulfill[t] == n) ? -(double)quant[t] : 0.0;
         // Warp-inclusive scan.
         for (int o = 1; o < 32; o <<= 1) {
-            float y = __shfl_up_sync(0xffffffffu, x, o);
+            double y = __shfl_up_sync(0xffffffffu, x, o);
             if (lane >= o) x += y;
         }
         if (lane == 31) wsum[warp] = x;
         __syncthreads();
         // Serial exclusive scan over the (≤32) per-warp totals.
         if (threadIdx.x == 0) {
-            float acc = 0.0f;
+            double acc = 0.0;
             for (int w = 0; w < nwarps; w++) { wexcl[w] = acc; acc += wsum[w]; }
             btot_s = acc;
         }
@@ -355,12 +355,12 @@ __global__ void capacity_cumsum_kernel(
 // output element (tid = (w·num_steps+s)·np1 + n); writes and the cumsum reads for a
 // warp (fixed slot, consecutive n) are coalesced. Mirrors the host gather exactly.
 __global__ void capacity_gather_kernel(
-    const float* __restrict__ cumsum,   // [eff, np1]
+    const double* __restrict__ cumsum,  // [eff, np1]
     const int* __restrict__ order_rel,  // [n_workers*num_steps], in [0,eff) or -1
     const int* __restrict__ fulfill,    // [eff]
     const int* __restrict__ quant,      // [eff]
     int n_workers, int num_steps, int eff, int np1,
-    float* __restrict__ cap_delta       // [n_workers*num_steps, np1]
+    double* __restrict__ cap_delta      // [n_workers*num_steps, np1]
 ) {
     long tid = blockIdx.x * (long)blockDim.x + threadIdx.x;
     long total = (long)n_workers * num_steps * np1;
@@ -370,13 +370,13 @@ __global__ void capacity_gather_kernel(
     int s  = (int)(ws % num_steps);
 
     int idx = order_rel[ws];
-    float cap_val = (idx >= 0 && idx < eff) ? cumsum[(size_t)idx * np1 + n] : 0.0f;
-    float cap_prev = 0.0f;
+    double cap_val = (idx >= 0 && idx < eff) ? cumsum[(size_t)idx * np1 + n] : 0.0;
+    double cap_prev = 0.0;
     if (s > 0) {
         int pidx = order_rel[ws - 1];
         if (pidx >= 0 && pidx < eff) cap_prev = cumsum[(size_t)pidx * np1 + n];
     }
-    float cd = cap_val - cap_prev;
-    float by_product = (idx >= 0 && idx < eff && fulfill[idx] == n) ? -(float)quant[idx] : 0.0f;
+    double cd = cap_val - cap_prev;
+    double by_product = (idx >= 0 && idx < eff && fulfill[idx] == n) ? -(double)quant[idx] : 0.0;
     cap_delta[tid] = -(cd - by_product);
 }
